feat(triggers): allow fractional refire cooldowns
Refire cooldown now accepts decimals (e.g. 1.5s) instead of integers
only. EQNag and EQLogParser exports commonly use sub-second cooldowns;
previously these were truncated on import (CooldownDuration int,
int(LockoutTime)), so 1.5 became 1. Widen RefireCooldownSecs to float64
end-to-end (model, API request, importers) and use step="any" +
parseFloat in the editor input.

SQLite stores the non-integer REAL losslessly in the INTEGER-affinity
column, so no migration is needed (covered by a new roundtrip test).

+++ backend/internal/api/triggers.go
@@ -62,7 +62,7 @@ type triggerRequest struct {
 	// the existing value (matching how cooldown_secs is left untouched). A
 	// present value — including 0 to clear it — replaces it. The editor always
 	// sends it; bulk paths (toggle, move category) omit it.
-	RefireCooldownSecs   *int                   `json:"refire_cooldown_secs,omitempty"`
+	RefireCooldownSecs   *float64               `json:"refire_cooldown_secs,omitempty"`
 	DisplayThresholdSecs int                    `json:"display_threshold_secs"`
 	BarColor             string                 `json:"bar_color"`
 	Characters           []string               `json:"characters"`
